+++ .cirrus/schema_test.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -euo pipefail
+
+cd "$(dirname "$0")/../tests/unit-test"
+
+go test -timeout=0 -v schema_test.go